# university/src/main/resources/openapi.v1.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Timetable",
    "version": "1.0",
    "contact": {
      "name": "Taras Shablii"
    },
    "description": "Timetable App API spec"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1/timetable"
    }
  ],
  "paths": {
...
    "/courses": {
      "get": {
        "summary": "List all courses",
        "tags": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Course"
                  }
                }
              }
            }
          }
        },
        "operationId": "get-courses",
        "description": "get all courses",
        "parameters": []
      },
      "post": {
        "summary": "Create new course",
        "operationId": "post-courses",
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "description": "create new course",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Course"
              }
            }
          }
        }
      }
    },
    "/courses/{id}": {
      "parameters": [
        {
          "schema": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          },
          "name": "id",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "Retrieve course",
        "tags": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "operationId": "get-courses-id",
        "description": "get single course by id"
      },
      "put": {
        "summary": "Edit existing course",
        "operationId": "put-courses-id",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "description": "edit course",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Course"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete course",
        "operationId": "delete-courses-id",
        "responses": {
          "501": {
            "description": "Not Implemented"
          }
        },
        "description": "delete course"
      }
    },
...
    "/professors/{id}/courses": {
      "parameters": [
        {
          "schema": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          },
          "name": "id",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "List all professor's courses",
        "tags": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Course"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "operationId": "get-professors-id-courses",
        "description": "get professor's courses"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Course": {
        "title": "Course",
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          },
          "name": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "name"
        ]
      },
...
    },
    "securitySchemes": {}
  }
}